added some select in sql for reviews model; started adding some session code for later;

+++ data/migrations/20200630095017_reviews.js
@@ -22,7 +22,7 @@ exports.up = function(knex) {
       .notNullable()
 
     table
-      .integer('rating')
+      .float('rating')
       .notNullable()
 
     table
